Database.Table.SQLite.Simple

Description

Storing Table in SQLite databases.

This module is meant to be imported qualified, i.e.

import qualified Database.Table.SQLite.Simple as Sql

This module provides a basic, self-contained set of functionalities for reading from and writing to SQLite databases. Specifically, we can

Convert between Haskell types and SQL types (via Database.SQLite.Simple).

Construct expressions, mainly for filter conditions.

Execute statements, such as retrieving and updating rows.

Open and close database files.

This module uses small embedded domain specific languages (eDSLs) to represent database operations — you do not need to write SQL source code. In other words, knowledge of SQL is not needed to use this module, but can be helpful.

Example

Here an example of how to use this module.

We begin by importing a helper module

import Data.Proxy ( Proxy (..) )

and the modules relating to databases:

import Database.Table
import qualified Database.Table.SQLite.Simple as Sql

Next, we define a type for the table that we want to store in the database:

type TablePerson =
    Table "person"
        :. Col "name" Text
        :. Col "birthyear" Int

The above type defines a table of name person with two columns name and birthyear. The type is automatically an instance of IsTableSql.

Here is an example row that we want to insert into the table:

exampleRow :: Row TablePerson
exampleRow = ("Ada Lovelace", 1815)

Next, we define two conveniences that bring the type level to the value level:

tablePerson :: Proxy TablePerson
tablePerson = Proxy

colBirthYear :: Col "birthyear" Int
colBirthYear = Col

Then, we define a sequence of operations that we want to perform on the database: First, we create the table, then we insert two rows, then, we ask for the number of rows that satisfy a certain condition:

example :: Sql.SqlM Int
example = do
    Sql.createTable tablePerson
    Sql.insertOne ("Babbage", 1791) tablePerson
    Sql.insertOne exampleRow tablePerson
    length <$> Sql.selectWhere (colBirthYear Sql.>. 1800) tablePerson

In the definition above, the type checker makes sure that

the inserted rows are valid for the table, and that

the value in the condition is valid for the column.

However, the type checker does not check that the condition mentions only columns that occur in the table.

Finally, we create an in-memory database, run the above operations and print the result. This result should be equal to 1, as the selectWhere operation should return exactly one row in our example.

main :: IO ()
main = do
    rowCount <- Sql.withConnection ":memory:"
        $ Sql.runSqlM example
    print rowCount -- should be equal to 1

Convert between Haskell types and SQL types

type IsTableSql t = (IsTable t, HasColumnsSql t, ToRow (Row t), FromRow (Row t))

Constaint for types that represent tables that can be mapped to SQL tables.

In particular, the type t tracks table names and column names (via IsTable), and has column types that can be converted to and from SQL types (via HasColumns, ToRow and FromRow).

Note: When using this constraint synonym, you will have to enable the FlexibleContexts extension.

Typically, types constructed using Table from Database.Table.Schema are automatically instances of IsTableSql, as long as their column types can be converted to and from SQL types. For example, the following type is automatically an instance:

type ExampleTable =
    Table "person"
        :. Col "name" Text
        :. Col "birthyear" Int

class HasColumnsSql t where

Class for types that correspond to a sequence (list) of SQL types.

Methods

getColumnSqlTypesSeq :: proxy t -> Seq SqlType

getColumnSqlTypes :: IsTableSql t => proxy t -> [SqlType]

Get column types as a plain list.

data SqlType

SQL column types, including constraints.

Internally represented as a textual string. Examples:

INTEGER  PRIMARY KEY NOT NULL
TEXT     NOT NULL

escapeSqlType :: SqlType -> Text

Print an SqlType in SQL syntax.

class (ToField a, FromField a) => IsColumnSql a where

Class that maps a Haskell type to a column type of an SQL database. Also ensures that the types can be converted to each other (via ToField and FromField from Database.SQLite.Simple).

Methods

getSqlType :: proxy a -> SqlType

SQL Expressions

data Expr a

SQL expressions corresponding to the Haskell type a.

true :: Expr Bool

SQL: true literal.

false :: Expr Bool

SQL: false literal.

not_ :: Expr Bool -> Expr Bool

SQL: NOT, negation.

(&&.) :: Expr Bool -> Expr Bool -> Expr Bool infixr 3

SQL: AND, conjunction.

and_ :: [Expr Bool] -> Expr Bool

Combine multiple Expr with (&&.).

(||.) :: Expr Bool -> Expr Bool -> Expr Bool infixr 2

SQL: OR, disjunction.

or_ :: [Expr Bool] -> Expr Bool

Combine multiple Expr with (||.).

(==.) :: forall (n :: Symbol) a. (IsColumnName n, ToField a) => Col n a -> a -> Expr Bool infix 4

SQL: ==, equality. Tests a column against a fixed value.

(/=.) :: forall (n :: Symbol) a. (IsColumnName n, ToField a) => Col n a -> a -> Expr Bool infix 4

SQL: !=, inequality. Tests a column against a fixed value.

(<.) :: forall (n :: Symbol) a. (IsColumnName n, ToField a, Ord a) => Col n a -> a -> Expr Bool infix 4

SQL: <, less than. Tests a column against a fixed value.

(<=.) :: forall (n :: Symbol) a. (IsColumnName n, ToField a, Ord a) => Col n a -> a -> Expr Bool infix 4

SQL: <=, less than or equal to. Tests a column against a fixed value.

(>.) :: forall (n :: Symbol) a. (IsColumnName n, ToField a, Ord a) => Col n a -> a -> Expr Bool infix 4

SQL: >, greater than. Tests a column against a fixed value.

(>=.) :: forall (n :: Symbol) a. (IsColumnName n, ToField a, Ord a) => Col n a -> a -> Expr Bool infix 4

SQL: >=, greater than or equal to. Tests a column against a fixed value.

SQL Statements

data SqlM a

Monad for database operations.

This monad includes the following effects:

mutable state — The state of the database is represented as mutable state.

exceptions

When using runSqlM, an action from this monad is executed atomically: Either the whole action succeeds, then all state changes persists, or the action throws an exception, then none of state changes persist.

data SqlException

Union of exceptions that can occur with Database.SQLite.Simple.

createTable :: IsTableSql t => proxy t -> SqlM ()

Create a new database table.

SQL: CREATE TABLE IF NOT EXISTS.

selectAll :: IsTableSql t => proxy t -> SqlM [Row t]

Retrieve all rows from the table.

SQL: SELECT ... FROM tablename.

selectWhere :: IsTableSql t => Expr Bool -> proxy t -> SqlM [Row t]

Retrieve those rows from the table that satisfy a condition.

SQL: SELECT ... FROM tablename WHERE condition.

insertOne :: IsTableSql t => Row t -> proxy t -> SqlM ()

Insert a single row into a database table.

SQL: INSERT INTO tablename VALUES ....

insertMany :: IsTableSql t => [Row t] -> proxy t -> SqlM ()

Insert many rows into a database table.

Uses a prepared SQL statement internally for efficiency.

SQL: INSERT INTO tablename VALUES ..., repeated.

deleteAll :: IsTableSql t => proxy t -> SqlM ()

Delete all rows from a database table.

SQL: DELETE FROM tablename.

deleteWhere :: IsTableSql t => Expr Bool -> proxy t -> SqlM ()

Delete those rows from a database table that satisfy a condition.

SQL: DELETE FROM tablename WHERE condition.

updateWhere

Arguments

:: IsTableSql t

=> Expr Bool
Condition.

-> [Update]
Updates to be performed.

-> proxy t
Table type.

-> SqlM ()

Update those rows in a database table that satisfy a condition.

SQL: UPDATE tablename SET ... WHERE condition.

data Update

Representation of a column update, for use with updateWhere.

Use (=.) to construct an Update.

(=.) :: forall (n :: Symbol) a. (IsColumnName n, ToField a) => Col n a -> a -> Update

Set the value of a given column.

SQL: SET columnName = value.

Open, run SqlM, and close databases

data Connection

A connection to an open SQLite database.

open :: FilePath -> IO Connection

Open a database connection to a given file. Will throw an exception if the connection cannot be opened, for example because the file does not exist.

Every open must be closed with a call to close.

If you specify ":memory:" or an empty string as the input filename, then a private, temporary in-memory database is created for the connection. This database will vanish when you close the connection.

(See also open in Database.SQLite.Simple.)

close :: Connection -> IO ()

Close a database connection.

(See also close in Database.SQLite.Simple.)

withConnection :: FilePath -> (Connection -> IO a) -> IO a

Opens a database connection, executes an action using this connection, and closes the connection, even in the presence of exceptions.

(See also withConnection in Database.SQLite.Simple.)

runSqlM :: SqlM a -> Connection -> IO a

Atomically run a computation from the SqlM monad.

exceptions — The change to the database state performed by SqlM completes either fully or not at all. If the computation throws an exception, then any intermediate state change is rolled back. (Essentially, the computation is wrapped in withTransaction.)

concurrency — A lock ensures that multiple calls to runSqlM are run in sequence.

rawSqlite :: (Connection -> IO a) -> SqlM a

Wrap a function from Database.SQLite.Simple in SqlM.