Database.Table.Schema

Description

The type Table represents the schema for a table in a relational database. We use it to

track table names and column names on the type level, and to

track column types as Haskell types.

More precisely, the schema for a table is represented by the combination of the types Table, (:.), and Col. Every table schema is an instance of the IsTable class.

Example: The following type is the schema for a database table with name person and two columns, name and birthyear, whose Haskell types are Text and Int, respectively.

type ExampleTable =
    Table "person"
        :. Col "name" Text
        :. Col "birthyear" Int

The type Table does not contain any data. Instead, use the Row type family to map a table schema to a type that contains one row worth of data.

Example: The following value is one potential row of a table that matches the table schema ExampleTable.

exampleRow :: Row ExampleTable
exampleRow = ("Ada Lovelace", 1815)

The types in this module are general, and not specific to any particular SQL implementation.

Table schema

class IsTable t where

Class of named tables with named columns.

The data contained in the table is essentially a list of rows with the given column names.

Methods

getTableName :: proxy t -> Text

getColumnNamesSeq :: proxy t -> Seq Text

getColumnNames :: IsTable t => proxy t -> [Text]

data Table (name :: Symbol)

Named database table.

Constructors

Table

data Col (name :: Symbol) a

Named database column.

Constructors

Col

data a :. b infixl 3

Infix notation for a pair of types.

Constructors

a :. b infixl 3

type IsColumnName (name :: Symbol) = KnownSymbol name

Constraint synonym for getColName.

getColumnName :: forall (name :: Symbol) a. IsColumnName name => Col name a -> Text

Get the name of a column from its type.

Table data: Row

type family Row t

Type family that maps a table schema t (which ideally satisfies IsTable t) to a type representing rows of that table.

newtype Only a

The 1-tuple type or single-value "collection".

This type is structurally equivalent to the Identity type, but its intent is more about serving as the anonymous 1-tuple type missing from Haskell for attaching typeclass instances.

Parameter usage example:

encodeSomething (Only (42::Int))

Result usage example:

xs <- decodeSomething
forM_ xs $ \(Only id) -> {- ... -}

Constructors

Only

Fields

fromOnly :: a